+++ agros-library/solver/problem.cpp
@@ -29,6 +29,7 @@
 
 #include "util/global.h"
 #include "util/constants.h"
+#include "util/conf.h"
 
 #include "field.h"
 #include "solutionstore.h"
@@ -702,7 +703,7 @@ bool ProblemBase::applyParametersInternal()
     // check materials
     foreach (SceneMaterial* material, m_scene->materials->items())
     {
-        foreach (ulong key, material->values().keys())
+        foreach (size_t key, material->values().keys())
         {
             if (!material->value(key)->isEvaluated())
             {
@@ -716,7 +717,7 @@ bool ProblemBase::applyParametersInternal()
     // check boundaries
     foreach (SceneBoundary* boundary, m_scene->boundaries->items())
     {
-        foreach (ulong key, boundary->values().keys())
+        foreach (size_t key, boundary->values().keys())
         {
             if (!boundary->value(key)->isEvaluated())
             {
@@ -1604,8 +1605,8 @@ void ProblemBase::writeProblemToJsonInternal(QJsonObject &rootJson)
             boundaryJson[BOUNDARY_FACES] = boundaryFacesJson;
 
             QJsonArray boundaryTypesJson;
-            const QMap<ulong, QSharedPointer<Value> > values = bound->values();
-            for (QMap<ulong, QSharedPointer<Value> >::const_iterator it = values.begin(); it != values.end(); ++it)
+            const QMap<size_t, QSharedPointer<Value> > values = bound->values();
+            for (QMap<size_t, QSharedPointer<Value> >::const_iterator it = values.begin(); it != values.end(); ++it)
             {
                 QJsonObject type;
                 type[ID] = bound->valueName(it.key());
@@ -1641,8 +1642,8 @@ void ProblemBase::writeProblemToJsonInternal(QJsonObject &rootJson)
             materialJson[MATERIAL_LABELS] = materialLabelsJson;
 
             QJsonArray materialTypesJson;
-            const QMap<ulong, QSharedPointer<Value> > values = mat->values();
-            for (QMap<ulong, QSharedPointer<Value> >::const_iterator it = values.begin(); it != values.end(); ++it)
+            const QMap<size_t, QSharedPointer<Value> > values = mat->values();
+            for (QMap<size_t, QSharedPointer<Value> >::const_iterator it = values.begin(); it != values.end(); ++it)
             {
                 QJsonObject type;
                 type[ID] = mat->valueName(it.key());
@@ -1772,7 +1773,7 @@ void Computation::readFromProblem()
                 bound->setValue(variable.id(), Value(this));
 
             // boundaries
-            foreach (ulong originValueIndex, originBoundary->values().keys())
+            foreach (size_t originValueIndex, originBoundary->values().keys())
             {
                 QSharedPointer<Value> originValue = originBoundary->value(originValueIndex);
 
@@ -1807,7 +1808,7 @@ void Computation::readFromProblem()
                 mat->setValue(variable.id(), Value(this));
 
             // materials
-            foreach (ulong originValueIndex, originMaterial->values().keys())
+            foreach (size_t originValueIndex, originMaterial->values().keys())
             {
                 QSharedPointer<Value> originValue = originMaterial->value(originValueIndex);
